Extract events mechanism into utils
