+++ src/events.ts
@@ -1,6 +1,7 @@
 import { Item } from "./items/item";
+import { MyEventSource } from "./lodash";
 
-type ItemEventDefinitions = {
+export type ItemEventDefinitions = {
   stateLoaded: Item;
 
   itemPlay: Item;
@@ -36,30 +37,6 @@ type ItemEventDefinitions = {
   "item.rightSidebarVisibilityChanged": Item;
 };
 
-export class AppEvents {
-  private itemCbs: Partial<Record<keyof ItemEventDefinitions, Action<Item>[]>> =
-    {};
+export type AppEvents = MyEventSource<ItemEventDefinitions>;
 
-  on<T extends keyof ItemEventDefinitions>(
-    type: T,
-    cb: Callback<ItemEventDefinitions[T]>
-  ) {
-    if (!this.itemCbs[type]) this.itemCbs[type] = [];
-    this.itemCbs[type]!.push(cb as any);
-  }
-
-  trigger<T extends keyof ItemEventDefinitions>(
-    ...args: TriggerArgs<T, ItemEventDefinitions[T]>
-  ) {
-    const [type, event] = args;
-    const cbs = this.itemCbs[type] || [];
-    cbs!.forEach((cb) => (cb as any)(event));
-  }
-}
-
-//Helper types for Events
-type Callback<T> = T extends void ? EmptyAction : Action<T>;
-
-type TriggerArgs<TKey, TVal> = TVal extends void
-  ? [type: TKey]
-  : [type: TKey, event: TVal];
+export const createAppEvents = () => new MyEventSource<ItemEventDefinitions>();
