+++ src/lodash/index.ts
@@ -1,1 +1,2 @@
 export * as array from "./array";
+export { MyEventSource } from "./events";
